Regressions and pointer arithmetic simplification and correction.

+++ src/include/t50_memalloc.h
@@ -25,7 +25,7 @@
 
 extern void *packet;
 
-void alloc_packet ( uint32_t );
+void alloc_packet ( size_t );
 void destroy_packet_buffer( void );
 
 #endif
