+++ src/include/t50_modules.h
@@ -21,6 +21,7 @@
 #ifndef __MODULES_INCLUDED__
 #define __MODULES_INCLUDED__
 
+#include <stddef.h>
 #include <stdint.h>
 #include <netinet/in.h>
 #include <t50_typedefs.h>
@@ -74,7 +75,7 @@ struct psdhdr
   uint16_t  len;        /* header length       */
 };
 
-typedef void ( *module_func_ptr_t ) ( const config_options_T * const restrict, uint32_t * restrict );
+typedef void ( *module_func_ptr_t ) ( const config_options_T * const restrict, size_t * restrict );
 
 /**
  * Modules entry structure.
@@ -109,19 +110,19 @@ void    build_indices ( void );
 uint32_t get_index ( config_options_T * );
 
 /* Modules functions prototypes. */
-void icmp ( const config_options_T * const restrict, uint32_t * restrict );
-void igmpv1 ( const config_options_T * const restrict, uint32_t * restrict );
-void igmpv3 ( const config_options_T * const restrict, uint32_t * restrict );
-void tcp ( const config_options_T * const restrict, uint32_t * restrict );
-void egp ( const config_options_T * const restrict, uint32_t * restrict );
-void udp ( const config_options_T * const restrict, uint32_t * restrict );
-void ripv1 ( const config_options_T * const restrict, uint32_t * restrict );
-void ripv2 ( const config_options_T * const restrict, uint32_t * restrict );
-void dccp ( const config_options_T * const restrict, uint32_t * restrict );
-void rsvp ( const config_options_T * const restrict, uint32_t * restrict );
-void ipsec ( const config_options_T * const restrict, uint32_t * restrict );
-void eigrp ( const config_options_T * const restrict, uint32_t * restrict );
-void ospf ( const config_options_T * const restrict, uint32_t * restrict );
+void icmp ( const config_options_T * const restrict, size_t * restrict );
+void igmpv1 ( const config_options_T * const restrict, size_t * restrict );
+void igmpv3 ( const config_options_T * const restrict, size_t * restrict );
+void tcp ( const config_options_T * const restrict, size_t * restrict );
+void egp ( const config_options_T * const restrict, size_t * restrict );
+void udp ( const config_options_T * const restrict, size_t * restrict );
+void ripv1 ( const config_options_T * const restrict, size_t * restrict );
+void ripv2 ( const config_options_T * const restrict, size_t * restrict );
+void dccp ( const config_options_T * const restrict, size_t * restrict );
+void rsvp ( const config_options_T * const restrict, size_t * restrict );
+void ipsec ( const config_options_T * const restrict, size_t * restrict );
+void eigrp ( const config_options_T * const restrict, size_t * restrict );
+void ospf ( const config_options_T * const restrict, size_t * restrict );
 /* --- add yours here */
 
 #endif
